requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
chromadb
sentence-transformers
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-chroma
langchain-huggingface==0.1.0